simple-cdd-meta/debian: Get rid of ssh static configuration
This is followup for recent change intended to remove postfix static
configuration. Motivation is the same as for postfix with addition
that sshd cryptography options availablility and defaults vary from
version to version due to security considerations and since one of
the project goals is to keep Debian system pristine, changing only
very limited set of parameters.

It is assumed that if later configuration reqired it should be performed
by third party automation tools like Ansible, Chief or Puppet.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>