simple-cdd-meta/debian: Skip all profile configs when loading config
This is micro optimization to speedup and keep from extra variables
being pulled to simple-cdd early during config load with --conf option
whose only indent is to provide list of $profiles for further
processing.

Now only $profiles, $auto_profiles and $simple_cdd_dir should be in
environment to be passed to simple-cdd on config load. They could be
unset(1) explicitly, but there is little reason in this as all of
these variables will be overwritten on each profile load.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>