simple-cdd-meta/debian: Fix Debian 8 (jessie) mirror settings
This is an oldoldstable release that receives little (or no) maintenance
and will be moved to archive.debian.org at some point.

However at this time it is still provided at deb.debian.org but without
updates and thus we should $updates_mirror should be disabled (set to
empty).

On the other hand $backports_mirror already provided only on
archive.debian.org, so we need to adjust this variable accordingly.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>