simple-cdd-meta/debian: Set profile variable to real profile name
Some profiles set 1st arg of include_deps_wrapper() to common profile
group name (e.g. locale and timezone) and thus setting $profile from
this argument isn't correct.

On the other hand we pass profile name assigned to __included_${1}__
variable that is used to stop recursion in profile inclusion. That is
passed as 3rd argument before this change. After this change we
reordered arguments to include_deps_wrapper() and other places to more
natural <group> <profile> <description> and thus it is passed on as 3d
argument.

While there make <description> really optional and make profile name
required argument that is assigned to __included_${1}__.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>