simple-cdd-meta/debian: Support for dynamic profile files generation
Motivation for this is to support single template file with preprocessor
conditionals like #if/#endif to define text blocks that should or should
not appear in final profile file based on certain condition like release
version or codename.

For example Debian packages might change their names between versions
like btrfs-tools for 9- and btrfs-progs for 10+, linux-tools for 9- and
linux-tools for 10+ or iproute for 9- and iproute2 for 10+, etc.

Another reason for this is to have different package sets for different
Debian versions as some packages might be unavailable for earlier
versions.

Note that "defs" file can't be put into $simple_cdd_temp directory as it
is defined in profile shell script (e.g. minimal.sh) and undefined for
rest profiles.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>