simple-cdd-meta/debian: Explicitly list xserver-xorg-{input,video}-all
Diskcheck complains about unsatisfied dependencies of xserver-xorg when
building on Debian GNU/Linux 9 (stretch) and above. It seems dependency
resolution process is broken somehow in reprepro(1) or there was changes
in package metadata (e.g. Depends tag points to Provides instead of
package name).

This issue however is not investigated carefully and for now with
dynamic profile files support we can eliminate most of such errors.

Fixes: 296ea3dc205b ("simple-cdd-meta/debian: Provide alternatives for package names")
Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>