simple-cdd-meta: Select correct packages for VA and VDPAU APIs
Packages that required for VA and VDPAU support depend on release:
use virtual packages when they available (e.g. va-driver-all on
Debian 8+ and vdpau-driver-all on Debian 9+) or their dependencies
when they are not available.

Explicitly list xserver-xorg-video-all and xserver-xorg-input-all
to make sure all video and input drivers added to image. Otherwise
xserver-xorg-video-nvidia might be selected and included to image
as OR dependency alternative in xorg-server package.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>