simple-cdd-meta/debian: Add option to enable systemd_cleanup_unit()
It is inappropriate to remove directives of unit file for recent
version of systemd as this may cause these units to malfunction.

Turn off this removal by default to avoid modifying unit files and
turn on for cases when such cleanup is really necessary (usually old
distros like Debian GNU/Linux 7.x (wheezy)).

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>