simple-cdd-meta/debian: Use POSIX shell compatible pattern negation
Negation in the following "${var##*[^a-z]*}" expression isn't POSIX
shell compatible because ^ as negation mark used in regular expressions,
while shell patternts aren't such ones.

Instead exclamation mark/point should be used for that purpose. Consider
below example code and results difference when running with ^ and ! as
negation:

  for neg in '^' '!'; do
      printf 'neg: %s\n' "$neg"

      for sh in 'bash' 'dash' 'mksh'; do
          /bin/$sh -c "
              app='abc'

              [ -n \"\${app##*[${neg}a-z]*}\" ] &&
                  v='v' || v='not-v'

              printf ' %-10s : %salid\n' \"\$0\" \"\$v\"
          "
      done
  done

  neg: ^
   /bin/bash  : valid
   /bin/dash  : not-valid
   /bin/mksh  : not-valid
  neg: !
   /bin/bash  : valid
   /bin/dash  : valid
   /bin/mksh  : valid

This clearly shows that non-bash shell interpreters does not support ^
and it should be replaced with ! to avoid bash(1) interpreter specific.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>